Azure_Code/Exercise-4-Azure-Function-Python/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
pandas==1.2.4
pyodbc==4.0.30
pymssql==2.2.1